storing fsm to db
