+++ scheduler/state_machine/controller.py
@@ -1,14 +1,19 @@
 import logging
-from celery import Celery
 from datetime import date, datetime, timedelta
-from state_machine import utils
+from state_machine.state_machine_utils import (compute_spent_weeks, create_new_date,
+                                               get_execution_week,
+                                               get_intervention_component, get_next_planned_date,
+                                               get_preferred_date_time, get_quit_date,
+                                               get_start_date,
+                                               is_new_week, plan_and_store, reschedule_dialog,
+                                               retrieve_intervention_day, schedule_next_execution,
+                                               store_completed_dialog, store_rescheduled_dialog,
+                                               store_scheduled_dialog, update_execution_week,
+                                               get_dialog_completion_state)
 from state_machine.const import (FUTURE_SELF_INTRO, GOAL_SETTING, TRACKING_DURATION,
-                                 PREPARATION_GA, MAX_PREPARATION_DURATION, EXECUTION_DURATION,
-                                 REDIS_URL, TRIGGER_COMPONENT, SCHEDULE_TRIGGER_COMPONENT)
+                                 PREPARATION_GA, MAX_PREPARATION_DURATION, EXECUTION_DURATION)
 from state_machine.state import State
 from virtual_coach_db.helper.definitions import (Components, Notifications)
-
-celery = Celery(broker=REDIS_URL)
 
 
 class OnboardingState(State):
@@ -22,9 +27,9 @@ class OnboardingState(State):
     def on_dialog_completed(self, dialog):
         logging.info('A dialog has been completed  %s ', dialog)
 
-        utils.store_completed_dialog(user_id=self.user_id,
-                                     dialog=dialog,
-                                     phase_id=1)
+        store_completed_dialog(user_id=self.user_id,
+                               dialog=dialog,
+                               phase_id=1)
 
         if dialog == Components.PREPARATION_INTRODUCTION:
             logging.info('Prep completed, starting profile creation')
@@ -74,11 +79,11 @@ class OnboardingState(State):
     def schedule_next_dialogs(self):
         # get the data on which the user has started the intervention
         # (day 1 of preparation phase)
-        start_date = utils.get_start_date(self.user_id)
+        start_date = get_start_date(self.user_id)
 
         # on day 9 at 10 a.m. send future self
-        fs_time = utils.create_new_date(start_date=start_date,
-                                        time_delta=FUTURE_SELF_INTRO)
+        fs_time = create_new_date(start_date=start_date,
+                                  time_delta=FUTURE_SELF_INTRO)
 
         plan_and_store(user_id=self.user_id,
                        dialog=Components.FUTURE_SELF_SHORT,
@@ -90,11 +95,11 @@ class OnboardingState(State):
         to day 9 of the preparation phase
         """
         first_date = date.today() + timedelta(days=1)
-        last_date = utils.get_start_date(self.user_id) + timedelta(days=8)
+        last_date = get_start_date(self.user_id) + timedelta(days=8)
 
         for day in range((last_date - first_date).days):
-            planned_date = utils.create_new_date(start_date=first_date,
-                                                 time_delta=day)
+            planned_date = create_new_date(start_date=first_date,
+                                           time_delta=day)
 
             plan_and_store(user_id=self.user_id,
                            dialog=Notifications.TRACK_NOTIFICATION,
@@ -107,20 +112,16 @@ class TrackingState(State):
         super().__init__(user_id)
         self.user_id = user_id
         self.state = State.TRACKING
-        # keep track of the completion of the
-        # future self dialog
-        self.self_completed = False
 
     def on_dialog_completed(self, dialog):
         logging.info('A dialog has been completed  %s ', dialog)
 
-        utils.store_completed_dialog(user_id=self.user_id,
-                                     dialog=dialog,
-                                     phase_id=1)
+        store_completed_dialog(user_id=self.user_id,
+                               dialog=dialog,
+                               phase_id=1)
 
         if dialog == Components.FUTURE_SELF_SHORT:
             logging.info('Future self completed')
-            self.self_completed = True
 
     def on_dialog_rescheduled(self, dialog, new_date):
 
@@ -144,12 +145,13 @@ class TrackingState(State):
 
         # if it's time and the self dialog has been completed,
         # move to new state
+        self_completed = get_dialog_completion_state(self.user_id, Components.FUTURE_SELF_SHORT)
         if (self.check_if_end_date(current_date) and
-                self.self_completed):
+                self_completed):
             self.set_new_state(GoalsSettingState(self.user_id))
 
     def check_if_end_date(self, date_to_check: date) -> bool:
-        intervention_day = utils.retrieve_intervention_day(self.user_id, date_to_check)
+        intervention_day = retrieve_intervention_day(self.user_id, date_to_check)
         # the Goal Setting state starts on day 10 of the intervention
         if intervention_day >= TRACKING_DURATION:
             return True
@@ -167,9 +169,9 @@ class GoalsSettingState(State):
     def on_dialog_completed(self, dialog):
         logging.info('A dialog has been completed  %s ', dialog)
 
-        utils.store_completed_dialog(user_id=self.user_id,
-                                     dialog=dialog,
-                                     phase_id=1)
+        store_completed_dialog(user_id=self.user_id,
+                               dialog=dialog,
+                               phase_id=1)
 
         if dialog == Components.GOAL_SETTING:
             logging.info('Goal setting completed, starting first aid kit')
@@ -197,18 +199,18 @@ class GoalsSettingState(State):
                        dialog=dialog)
 
     def plan_buffer_phase_dialogs(self):
-        quit_date = utils.get_quit_date(self.user_id)
-        start_date = utils.get_start_date(self.user_id)
+        quit_date = get_quit_date(self.user_id)
+        start_date = get_start_date(self.user_id)
 
         if (quit_date - start_date).days >= PREPARATION_GA:
-            planned_date = utils.create_new_date(start_date=start_date, time_delta=PREPARATION_GA)
+            planned_date = create_new_date(start_date=start_date, time_delta=PREPARATION_GA)
             plan_and_store(user_id=self.user_id,
                            dialog=Components.GENERAL_ACTIVITY,
                            planned_date=planned_date)
 
         if (quit_date - start_date).days == MAX_PREPARATION_DURATION:
-            planned_date = utils.create_new_date(start_date=start_date,
-                                                 time_delta=MAX_PREPARATION_DURATION)
+            planned_date = create_new_date(start_date=start_date,
+                                           time_delta=MAX_PREPARATION_DURATION)
             plan_and_store(user_id=self.user_id,
                            dialog=Components.GENERAL_ACTIVITY,
                            planned_date=planned_date)
@@ -216,8 +218,8 @@ class GoalsSettingState(State):
     def plan_execution_start_dialog(self):
         # this is the intro video to be sent the first time
         # the execution starts (not after lapse/relapse)
-        quit_date = utils.get_quit_date(self.user_id)
-        planned_date = utils.create_new_date(start_date=quit_date)
+        quit_date = get_quit_date(self.user_id)
+        planned_date = create_new_date(start_date=quit_date)
 
         plan_and_store(user_id=self.user_id,
                        dialog=Components.EXECUTION_INTRODUCTION,
@@ -230,14 +232,14 @@ class GoalsSettingState(State):
         first_date = date.today() + timedelta(days=1)
 
         # the time span to quit date
-        buffer_length = (utils.get_quit_date(self.user_id) - first_date).days
+        buffer_length = (get_quit_date(self.user_id) - first_date).days
 
         total_duration = buffer_length + EXECUTION_DURATION
         last_date = first_date + timedelta(days=total_duration)
 
         for day in range((last_date - first_date).days):
-            planned_date = utils.create_new_date(start_date=first_date,
-                                                 time_delta=day)
+            planned_date = create_new_date(start_date=first_date,
+                                           time_delta=day)
 
             plan_and_store(user_id=self.user_id,
                            dialog=Notifications.PA_NOTIFICATION,
@@ -245,7 +247,7 @@ class GoalsSettingState(State):
 
     def run(self):
 
-        start_date = utils.get_start_date(self.user_id)
+        start_date = get_start_date(self.user_id)
 
         # if the starting of this phase is after the expected one
         # launch immediately the goal setting
@@ -254,8 +256,8 @@ class GoalsSettingState(State):
 
         else:
             # on day 10 at 10 a.m. send future self plan goal setting
-            gs_time = utils.create_new_date(start_date=start_date,
-                                            time_delta=GOAL_SETTING)
+            gs_time = create_new_date(start_date=start_date,
+                                      time_delta=GOAL_SETTING)
 
         plan_and_store(user_id=self.user_id,
                        dialog=Components.GOAL_SETTING,
@@ -284,7 +286,7 @@ class BufferState(State):
                        dialog=dialog)
 
     def check_if_end_date(self, current_date: datetime.date):
-        quit_date = utils.get_quit_date(self.user_id)
+        quit_date = get_quit_date(self.user_id)
 
         if current_date >= quit_date:
             logging.info('Buffer sate ended, starting execution state')
@@ -301,9 +303,9 @@ class ExecutionRunState(State):
     def on_dialog_completed(self, dialog):
         logging.info('A dialog has been completed  %s ', dialog)
 
-        utils.store_completed_dialog(user_id=self.user_id,
-                                     dialog=dialog,
-                                     phase_id=2)
+        store_completed_dialog(user_id=self.user_id,
+                               dialog=dialog,
+                               phase_id=2)
 
         if dialog == Components.EXECUTION_INTRODUCTION:
             logging.info('Execution intro completed, starting general activity')
@@ -318,7 +320,7 @@ class ExecutionRunState(State):
         elif dialog == Components.WEEKLY_REFLECTION:
             logging.info('Weekly reflection completed')
 
-            week = utils.get_execution_week(user_id=self.user_id)
+            week = get_execution_week(user_id=self.user_id)
 
             # if in week 3 or 8 of the execution, run future self after
             # completing the weekly reflection
@@ -358,17 +360,17 @@ class ExecutionRunState(State):
 
     def on_new_day(self, current_date: datetime.date):
 
-        quit_date = utils.get_quit_date(self.user_id)
+        quit_date = get_quit_date(self.user_id)
 
         # in case the current day of the week is the same as the one set in the
         # quit_date (and is not the same date), a new week started.
         # Thus, the execution week must be updated
-        if utils.is_new_week(current_date, quit_date):
+        if is_new_week(current_date, quit_date):
             # get the current week number
-            week_number = utils.compute_spent_weeks(current_date, quit_date)
+            week_number = compute_spent_weeks(current_date, quit_date)
 
             # increase the week number by one
-            utils.update_execution_week(self.user_id, week_number)
+            update_execution_week(self.user_id, week_number)
 
     def run(self):
         logging.info("Running state %s", self.state)
@@ -387,9 +389,9 @@ class RelapseState(State):
     def on_dialog_completed(self, dialog):
         logging.info('A dialog has been completed  %s ', dialog)
 
-        utils.store_completed_dialog(user_id=self.user_id,
-                                     dialog=dialog,
-                                     phase_id=3)
+        store_completed_dialog(user_id=self.user_id,
+                               dialog=dialog,
+                               phase_id=3)
 
         if dialog in [Components.RELAPSE_DIALOG,
                       Components.RELAPSE_DIALOG_HRS,
@@ -399,7 +401,7 @@ class RelapseState(State):
 
             logging.info('Relapse dialog completed ')
 
-            quit_date = utils.get_quit_date(self.user_id)
+            quit_date = get_quit_date(self.user_id)
             current_date = date.today()
 
             # if the quit date is in the future, it has been reset
@@ -442,14 +444,14 @@ class ClosingState(State):
         logging.info("Running state %s", self.state)
         # plan the execution of the closing dialog
 
-        component = utils.get_intervention_component(Components.CLOSING_DIALOG)
+        component = get_intervention_component(Components.CLOSING_DIALOG)
 
-        closing_date = utils.get_preferred_date_time(
+        closing_date = get_next_planned_date(
             user_id=self.user_id,
             intervention_component_id=component.intervention_component_id
-        )[0]
+        )
 
-        planned_date = utils.create_new_date(closing_date)
+        planned_date = create_new_date(closing_date)
 
         plan_and_store(user_id=self.user_id,
                        dialog=Components.CLOSING_DIALOG,
@@ -462,98 +464,15 @@ class ClosingState(State):
     def on_dialog_completed(self, dialog):
         logging.info('A dialog has been completed  %s ', dialog)
 
-        utils.store_completed_dialog(user_id=self.user_id,
-                                     dialog=dialog,
-                                     phase_id=2)
+        store_completed_dialog(user_id=self.user_id,
+                               dialog=dialog,
+                               phase_id=2)
 
         if dialog == Components.CLOSING_DIALOG:
             logging.info('Closing dialog completed. Intervention finished')
 
     def on_dialog_rescheduled(self, dialog, new_date):
-
         reschedule_dialog(user_id=self.user_id,
                           dialog=dialog,
                           planned_date=new_date,
                           phase=1)
-
-
-def plan_and_store(user_id: int, dialog: str, planned_date: datetime = None):
-    """
-    Program a celery task for the planned_date, or sends it immediately in case
-    planned_date is None, and stores the new component to the DB
-    Args:
-        user_id:user id
-        dialog: dialog to be triggered
-        planned_date: date when the dialog has to be triggered
-    Returns:
-
-    """
-
-    component = utils.get_intervention_component(dialog)
-    dialog_id = component.intervention_component_id
-    trigger = component.intervention_component_trigger
-
-    if planned_date is None:
-        celery.send_task(TRIGGER_COMPONENT, (user_id, trigger))
-
-        utils.store_scheduled_dialog(user_id=user_id,
-                                     dialog_id=dialog_id,
-                                     phase_id=1)
-    else:
-        task = celery.send_task(SCHEDULE_TRIGGER_COMPONENT,
-                                (user_id, trigger),
-                                eta=planned_date)
-
-        utils.store_scheduled_dialog(user_id=user_id,
-                                     dialog_id=dialog_id,
-                                     phase_id=1,
-                                     planned_date=planned_date,
-                                     task_uuid=str(task.task_id))
-
-
-def reschedule_dialog(user_id: int, dialog: str, planned_date: datetime, phase: int):
-    """
-    Program a new celery task for the planned_date and store the info in the db
-    Args:
-        user_id:user id
-        dialog: dialog to be triggered
-        planned_date: date when the dialog has to be triggered
-        phase: db id of the phase
-    Returns:
-
-    """
-    component = utils.get_intervention_component(dialog)
-    dialog_id = component.intervention_component_id
-    trigger = component.intervention_component_trigger
-
-    task = celery.send_task(SCHEDULE_TRIGGER_COMPONENT,
-                            (user_id, trigger),
-                            eta=planned_date)
-
-    utils.store_rescheduled_dialog(user_id=user_id,
-                                   dialog_id=dialog_id,
-                                   phase_id=phase,
-                                   planned_date=planned_date,
-                                   task_uuid=str(task.task_id))
-
-
-def schedule_next_execution(user_id: int, dialog: str):
-    """
-    Get the next expected execution date for an intervention component,
-    and schedules it
-    Args:
-        user_id: id of the user
-        dialog: Name of the component
-
-    """
-    component = utils.get_intervention_component(dialog)
-    component_id = component.intervention_component_id
-
-    planned_date = utils.get_next_planned_date(user_id=user_id,
-                                               intervention_component_id=component_id)
-
-    new_date = planned_date.replace(hour=10, minute=00)
-
-    plan_and_store(user_id=user_id,
-                   dialog=dialog,
-                   planned_date=new_date)
